rename package structuringprograms to gettingstarted
