+++ answerkey/gettingstarted/3.answer.scala
