+++ answerkey/gettingstarted/6.answer.scala
