+++ answerkey/gettingstarted/6.hint.txt
