+++ answerkey/gettingstarted/7.answer.scala
